jewelry/requirements.txt
chromadb~=0.3.29
langchain_community
langchain_openai
langchain_huggingface
langchain-text-splitters
git+https://github.com/mlrun/genai-factory.git
urllib3==1.26.5